update server, update routes error filer livro.ts, update new folder migration
